Add collision layers
Added collision layers.
'CollisionLayers' is the layers the object exists in (in binary). 'CoupleWithLayers' is the layers the object interacts with: for example, if the floor has CollisionLayers = 0b01 and the player has CoupleWithLayers = 0b11, the player can attempt collisions with the floor - but if CoupleWithLayers = 0b10, this is not the case.